Updated List and Map to use generics.

+++ src/main/java/com/l2fprod/common/swing/renderer/ColorCellRenderer.java
@@ -68,7 +68,7 @@ public class ColorCellRenderer extends DefaultCellRenderer {
             return null;
         }
         if (value instanceof Integer) {
-            value = new Color(((Integer) value).intValue());
+            value = new Color(((Integer) value));
         }
         return new PaintIcon((Paint) value);
     }
